consolidate deserializers
