+++ client/core/rs/src/entities/mod.rs
@@ -10,17 +10,17 @@ use clap::Parser;
 use derive_empty_traits::EmptyTraits;
 use derive_variants::{EnumVariants, ExtractVariant};
 use serde::{
-  de::{
-    value::{MapAccessDeserializer, SeqAccessDeserializer},
-    Visitor,
-  },
-  Deserialize, Deserializer, Serialize,
+  de::{value::MapAccessDeserializer, Visitor},
+  Deserialize, Serialize,
 };
 use serror::Serror;
 use strum::{AsRefStr, Display, EnumString};
 use typeshare::typeshare;
 
-use crate::parser::parse_key_value_list;
+use crate::{
+  deserializers::file_contents_deserializer,
+  parsers::parse_key_value_list,
+};
 
 /// Subtypes of [Alert][alert::Alert].
 pub mod alert;
@@ -363,210 +363,6 @@ pub fn environment_vars_from_str(
       .map(|(variable, value)| EnvironmentVar { variable, value })
       .collect()
   })
-}
-
-pub fn env_vars_deserializer<'de, D>(
-  deserializer: D,
-) -> Result<String, D::Error>
-where
-  D: Deserializer<'de>,
-{
-  deserializer.deserialize_any(EnvironmentVarVisitor)
-}
-
-pub fn option_env_vars_deserializer<'de, D>(
-  deserializer: D,
-) -> Result<Option<String>, D::Error>
-where
-  D: Deserializer<'de>,
-{
-  deserializer.deserialize_any(OptionEnvVarVisitor)
-}
-
-struct EnvironmentVarVisitor;
-
-impl<'de> Visitor<'de> for EnvironmentVarVisitor {
-  type Value = String;
-
-  fn expecting(
-    &self,
-    formatter: &mut std::fmt::Formatter,
-  ) -> std::fmt::Result {
-    write!(formatter, "string or Vec<EnvironmentVar>")
-  }
-
-  fn visit_str<E>(self, v: &str) -> Result<Self::Value, E>
-  where
-    E: serde::de::Error,
-  {
-    let out = v.to_string();
-    if out.is_empty() || out.ends_with('\n') {
-      Ok(out)
-    } else {
-      Ok(out + "\n")
-    }
-  }
-
-  fn visit_seq<A>(self, seq: A) -> Result<Self::Value, A::Error>
-  where
-    A: serde::de::SeqAccess<'de>,
-  {
-    let vars = Vec::<EnvironmentVar>::deserialize(
-      SeqAccessDeserializer::new(seq),
-    )?;
-    let vars = vars
-      .iter()
-      .map(|EnvironmentVar { variable, value }| {
-        format!("  {variable} = {value}")
-      })
-      .collect::<Vec<_>>()
-      .join("\n");
-    let extra = if vars.is_empty() { "" } else { "\n" };
-    Ok(vars + extra)
-  }
-}
-
-struct OptionEnvVarVisitor;
-
-impl<'de> Visitor<'de> for OptionEnvVarVisitor {
-  type Value = Option<String>;
-
-  fn expecting(
-    &self,
-    formatter: &mut std::fmt::Formatter,
-  ) -> std::fmt::Result {
-    write!(formatter, "null or string or Vec<EnvironmentVar>")
-  }
-
-  fn visit_str<E>(self, v: &str) -> Result<Self::Value, E>
-  where
-    E: serde::de::Error,
-  {
-    EnvironmentVarVisitor.visit_str(v).map(Some)
-  }
-
-  fn visit_seq<A>(self, seq: A) -> Result<Self::Value, A::Error>
-  where
-    A: serde::de::SeqAccess<'de>,
-  {
-    EnvironmentVarVisitor.visit_seq(seq).map(Some)
-  }
-
-  fn visit_none<E>(self) -> Result<Self::Value, E>
-  where
-    E: serde::de::Error,
-  {
-    Ok(None)
-  }
-
-  fn visit_unit<E>(self) -> Result<Self::Value, E>
-  where
-    E: serde::de::Error,
-  {
-    Ok(None)
-  }
-}
-
-pub fn labels_deserializer<'de, D>(
-  deserializer: D,
-) -> Result<String, D::Error>
-where
-  D: Deserializer<'de>,
-{
-  deserializer.deserialize_any(LabelVisitor)
-}
-
-pub fn option_labels_deserializer<'de, D>(
-  deserializer: D,
-) -> Result<Option<String>, D::Error>
-where
-  D: Deserializer<'de>,
-{
-  deserializer.deserialize_any(OptionLabelVisitor)
-}
-
-struct LabelVisitor;
-
-impl<'de> Visitor<'de> for LabelVisitor {
-  type Value = String;
-
-  fn expecting(
-    &self,
-    formatter: &mut std::fmt::Formatter,
-  ) -> std::fmt::Result {
-    write!(formatter, "string or Vec<EnvironmentVar>")
-  }
-
-  fn visit_str<E>(self, v: &str) -> Result<Self::Value, E>
-  where
-    E: serde::de::Error,
-  {
-    let out = v.to_string();
-    if out.is_empty() || out.ends_with('\n') {
-      Ok(out)
-    } else {
-      Ok(out + "\n")
-    }
-  }
-
-  fn visit_seq<A>(self, seq: A) -> Result<Self::Value, A::Error>
-  where
-    A: serde::de::SeqAccess<'de>,
-  {
-    let vars = Vec::<EnvironmentVar>::deserialize(
-      SeqAccessDeserializer::new(seq),
-    )?;
-    let vars = vars
-      .iter()
-      .map(|EnvironmentVar { variable, value }| {
-        format!("  {variable}: {value}")
-      })
-      .collect::<Vec<_>>()
-      .join("\n");
-    let extra = if vars.is_empty() { "" } else { "\n" };
-    Ok(vars + extra)
-  }
-}
-
-struct OptionLabelVisitor;
-
-impl<'de> Visitor<'de> for OptionLabelVisitor {
-  type Value = Option<String>;
-
-  fn expecting(
-    &self,
-    formatter: &mut std::fmt::Formatter,
-  ) -> std::fmt::Result {
-    write!(formatter, "null or string or Vec<EnvironmentVar>")
-  }
-
-  fn visit_str<E>(self, v: &str) -> Result<Self::Value, E>
-  where
-    E: serde::de::Error,
-  {
-    LabelVisitor.visit_str(v).map(Some)
-  }
-
-  fn visit_seq<A>(self, seq: A) -> Result<Self::Value, A::Error>
-  where
-    A: serde::de::SeqAccess<'de>,
-  {
-    LabelVisitor.visit_seq(seq).map(Some)
-  }
-
-  fn visit_none<E>(self) -> Result<Self::Value, E>
-  where
-    E: serde::de::Error,
-  {
-    Ok(None)
-  }
-
-  fn visit_unit<E>(self) -> Result<Self::Value, E>
-  where
-    E: serde::de::Error,
-  {
-    Ok(None)
-  }
 }
 
 #[typeshare]
@@ -1166,84 +962,5 @@ impl ResourceTargetVariant {
       ResourceTargetVariant::ResourceSync => "resource_sync",
       ResourceTargetVariant::Stack => "stack",
     }
-  }
-}
-
-/// Using this ensures the file contents end with trailing '\n'
-pub fn file_contents_deserializer<'de, D>(
-  deserializer: D,
-) -> Result<String, D::Error>
-where
-  D: Deserializer<'de>,
-{
-  deserializer.deserialize_any(FileContentsVisitor)
-}
-
-/// Using this ensures the file contents end with trailing '\n'
-pub fn option_file_contents_deserializer<'de, D>(
-  deserializer: D,
-) -> Result<Option<String>, D::Error>
-where
-  D: Deserializer<'de>,
-{
-  deserializer.deserialize_any(OptionFileContentsVisitor)
-}
-
-struct FileContentsVisitor;
-
-impl<'de> Visitor<'de> for FileContentsVisitor {
-  type Value = String;
-
-  fn expecting(
-    &self,
-    formatter: &mut std::fmt::Formatter,
-  ) -> std::fmt::Result {
-    write!(formatter, "string")
-  }
-
-  fn visit_str<E>(self, v: &str) -> Result<Self::Value, E>
-  where
-    E: serde::de::Error,
-  {
-    let out = v.trim_end().to_string();
-    if out.is_empty() {
-      Ok(out)
-    } else {
-      Ok(out + "\n")
-    }
-  }
-}
-
-struct OptionFileContentsVisitor;
-
-impl<'de> Visitor<'de> for OptionFileContentsVisitor {
-  type Value = Option<String>;
-
-  fn expecting(
-    &self,
-    formatter: &mut std::fmt::Formatter,
-  ) -> std::fmt::Result {
-    write!(formatter, "null or string")
-  }
-
-  fn visit_str<E>(self, v: &str) -> Result<Self::Value, E>
-  where
-    E: serde::de::Error,
-  {
-    FileContentsVisitor.visit_str(v).map(Some)
-  }
-
-  fn visit_none<E>(self) -> Result<Self::Value, E>
-  where
-    E: serde::de::Error,
-  {
-    Ok(None)
-  }
-
-  fn visit_unit<E>(self) -> Result<Self::Value, E>
-  where
-    E: serde::de::Error,
-  {
-    Ok(None)
   }
 }
